Fix issue 4: Implement discrete grid movement
The game now only checks for input when the Vehicle is in the
center of a tile. Well, that was the idea, but it actually checks
every half tile, so the Vehicle can move along the edge of a tile
too. I might actually keep this and allow the player to drive
down the middle of the road. We'll see. Also added some
documentation.