Check Tile for 'road' type before allowing input
Had to change the coordinate system to start at 0 instead of 1.
Otherwise I would have had to do a little extra work to access
the Tile array, and who wants to do that. I will need to fix the
grid movement to properly restrict it to whole tiles, instead of
allowing cars to drive on the edge. Right now it doesn't let the
Vehicle drive in the middle of road tiles with land north of them.