TEMP Fix issue 8: Smooth movement between tiles
The 'fix' is just to move the car a set number of pixels each
frame. Assuming VSync, 60fps, and stable perfomance this is a
working fix, but that's a lot of assumptions. In the long run
not a reasonable solution, but it will allow me to make some
progress in other areas of the game.