Fix issue 9: Implement car-like controls (buggy)
The car is now controlled with API methods that direct movement
relative to the car instead of relative to the world. This is a
little cleaner since the game doesn't have to directly set the
velocity of the car. However shifting input to be from the car's
perspective does break the existing system to keep the car from
driving off the road.