Add checks with tile neighbors to keep car on road
There are two bugs present:
If a car cannot turn, it doesn't then check to see if it can still
go forward, allowing the player to drive of the edge of the map.
The other bug is a little unclear. The tile successors don't seem
to be working on the East and West edges of the map