Add logic for two levels
Doing this the morning of, so it's kinda just slapped together.
A new variable in the game class tracks the level, and the game
states check the level to act apporpriately. After fininshing the
first level, the game goes back to the launch screen. When
PlayState is re-entered, Level 2 is drawn.